requirements.txt
transformers==4.57.3
datasets==4.4.1
hydra-core==1.3.2
wandb==0.23.0
